+++ apps/nextjs/src/app/page.tsx
@@ -1,6 +1,8 @@
 import { Suspense } from "react";
 
-import { AuthShowcase } from "./auth-showcase";
+import { auth } from "@acme/auth";
+
+import { SignIn, SignOut } from "~/components/auth";
 import { CreatePostForm, PostList } from "./posts";
 
 export default function HomePage() {
@@ -18,5 +20,32 @@ export default function HomePage() {
         </Suspense>
       </div>
     </main>
+  );
+}
+
+export async function AuthShowcase() {
+  const session = await auth();
+
+  if (!session) {
+    return (
+      <SignIn
+        provider="google"
+        className="rounded-full bg-white/10 px-10 py-3 font-semibold text-white no-underline transition hover:bg-white/20"
+      >
+        Sign in with Discord
+      </SignIn>
+    );
+  }
+
+  return (
+    <div className="flex flex-col items-center justify-center gap-4">
+      <p className="text-center text-2xl text-white">
+        {session && <span>Logged in as {session.user.name}</span>}
+      </p>
+
+      <SignOut className="rounded-full bg-white/10 px-10 py-3 font-semibold text-white no-underline transition hover:bg-white/20">
+        Sign out
+      </SignOut>
+    </div>
   );
 }
