+++ apps/nextjs/src/components/auth.tsx
@@ -1,0 +1,24 @@
+import type { ComponentProps } from "react";
+
+import { CSRF_experimental } from "@acme/auth";
+
+export function SignIn({
+  provider,
+  ...props
+}: { provider: string } & ComponentProps<"button">) {
+  return (
+    <form action={`/api/auth/signin/${provider}`} method="post">
+      <button {...props} />
+      <CSRF_experimental />
+    </form>
+  );
+}
+
+export function SignOut(props: ComponentProps<"button">) {
+  return (
+    <form action="/api/auth/signout" method="post">
+      <button {...props} />
+      <CSRF_experimental />
+    </form>
+  );
+}
